Windows: escape arguments in daemon instead of libcontainerd
The appropriate behavior for process arguments in Windows is really to
join the arguments with spaces since Windows doesn't really have the
notion of an array of arguments, just an arguments string. As a result,
any automatic escaping in libcontainerd would prohibit certain legitmiate
command lines.

This change moves the escaping that was in libcontainerd back into daemon.
This also returns the behavior of RUN in the builder to the pre-containerd
behavior. This may change again in the future, but it should be separate
from the containerd integration.

Signed-off-by: John Starks <jostarks@microsoft.com>